app/requirements.txt
streamlit
opencv-python
pyzbar
Pillow
numpy
pandas
XlsxWriter
